Add factorizations that prove Conjecture 2 in the range [43632, 44760].
Conjecture 2 says that t(n) >= n/3 for n >= 300_000.  One thing we are
interested in is how small can the lower threshold 300_000 be?

I believe the minimal threshold is exactly 43632, that is t(n) >= n/3
for all n >= 43632, while t(n) < n/3 for n = 43631.

The factorizations included in this commit witness that the conjecture
is true in the range 43632 <= n <= 44760.  Specifically, this is done
by including files such as 43632-43632-14545.txt which witnesses
t(43632) >= 14545.  All of them have been checked using
simple_check.py , which takes about one second per file or two minutes
total to run.

Almost all of these factorizations are probably optimal.  I have some
uncertainty about two: t(43885) >= 14633 and t(43987) >= 14664 might
not be tight.  However, there is no proof included here that the
factorizations are optimal (nor that one isn't possible for n=43631),
and I don't know how to produce one yet.

An effort was made to use as few witnesses as possible.  For example,
t(43632) >= 14545 implies that t(n) >= n/3 for n in the range [43632,
43635], so the next number we consider is the smallest missing: 43636.

It should be possible to extend these witnesses up to 80000 or more,
which would meet up with existing results.