Add factorizations that prove Conjecture 2 in the range [43632, 80973].
This builds on top of the previous commit that did the range [43632,
44760].

This links up with the previous work that proves Conjecture 2 in a
range starting at 80000.

Some computational difficulties during this run means that some of the
files here are not necessarily maximal factorizations, though most
should still be.